+++ sdk/communication/azure-resourcemanager-communication/CHANGELOG.md
@@ -1,14 +1,43 @@
 # Release History
 
-## 2.3.0-beta.1 (Unreleased)
+## 1.0.0-beta.1 (2025-01-24)
+
+- Azure Resource Manager Communication client library for Java. This package contains Microsoft Azure SDK for Communication Management SDK. REST API for Azure Communication Services. Package tag package-preview-2023-06. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Features Added
 
-### Breaking Changes
+* `models.SuppressionListAddressResourceCollection` was added
 
-### Bugs Fixed
+* `models.SuppressionListAddressResource` was added
 
-### Other Changes
+* `models.SuppressionListResource$UpdateStages` was added
+
+* `models.SuppressionListResource$Update` was added
+
+* `models.SuppressionListAddressResource$Update` was added
+
+* `models.SuppressionListAddressResource$Definition` was added
+
+* `models.SuppressionListAddresses` was added
+
+* `models.SuppressionListAddressResource$DefinitionStages` was added
+
+* `models.SuppressionListResource$Definition` was added
+
+* `models.SuppressionListResource` was added
+
+* `models.SuppressionLists` was added
+
+* `models.SuppressionListResourceCollection` was added
+
+* `models.SuppressionListResource$DefinitionStages` was added
+
+* `models.SuppressionListAddressResource$UpdateStages` was added
+
+#### `CommunicationManager` was modified
+
+* `suppressionListAddresses()` was added
+* `suppressionLists()` was added
 
 ## 2.2.0 (2024-12-19)